+++ lib/mininet.py
@@ -11,6 +11,10 @@ import network
 
 class MiniNet:
 
+    def __init__(self):
+        print('INFO:  WiFi STA: Starting MiniNet')
+        self.station = network.WLAN()
+
     def connect_wifi(self, ssid, password, timeout=3000):
         """
         https://docs.pycom.io/firmwareapi/pycom/network/wlan/
@@ -26,20 +30,25 @@ class MiniNet:
         #server.deinit()
 
         # Connect to WLAN
-        print('INFO:  WiFi STA: Starting MiniNet')
-        wlan = network.WLAN(mode=network.WLAN.STA)
-        wlan.disconnect()
-        nets = wlan.scan()
+        print('INFO:  WiFi STA: Scanning for networks')
+        #self.station = network.WLAN(mode=network.WLAN.STA)
+        try:
+            self.station.disconnect()
+        except:
+            pass
+
+        self.station.mode(network.WLAN.STA)
+        nets = self.station.scan()
         ssids = [net.ssid for net in nets]
 
         print('INFO:  WiFi STA: Networks found {}'.format(ssids))
         for net in nets:
             if net.ssid == ssid:
                 print('INFO:  WiFi STA: Connecting to "{}"'.format(ssid))
-                wlan.connect(net.ssid, auth=(net.sec, password), timeout=timeout)
+                self.station.connect(net.ssid, auth=(net.sec, password), timeout=timeout)
 
                 try:
-                    while not wlan.isconnected():
+                    while not self.station.isconnected():
                         # Save power while waiting
                         machine.idle()
                         time.sleep_ms(250)
@@ -61,9 +70,11 @@ class MiniNet:
 
         #self.wait_for_nic()
 
-        if wlan.isconnected():
-            print('INFO:  WiFi STA: Network configuration is {}'.format(wlan.ifconfig()))
+        if self.station.isconnected():
+            print('INFO:  WiFi STA: Network configuration is {}'.format(self.station.ifconfig()))
             print('INFO:  Ready.')
+            print()
+            print('INFO:  WiFi STA: IP address is {}'.format(self.get_ip_address()))
 
         print()
         print('Note: Press CTRL+X or Ctrl+] to detach from the REPL')
@@ -82,3 +93,6 @@ class MiniNet:
             time.sleep(0.25)
             attempts += 1
         print('Network interface ready')
+
+    def get_ip_address(self):
+        return self.station.ifconfig()[0]